Added ResponseEntity to the Student Controller

+++ src/main/java/com/example/SpringRESTAPILBD/student/StudentController.java
@@ -1,6 +1,8 @@
 package com.example.SpringRESTAPILBD.student;
 
+import com.example.SpringRESTAPILBD.SpringRestApiLbdApplication;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,27 +18,48 @@ public class StudentController {
     }
 
     @GetMapping("/all")
-    public List<Student> getAllStudents() {
-        return studentService.getAllStudents();
+    ResponseEntity<String> getAllStudents() {
+        return ResponseEntity.ok().header("Successful:", "true")
+                .body(SpringRestApiLbdApplication.asJsonString(studentService.getAllStudents()));
     }
 
     @PostMapping
-    public void addStudent(@RequestBody Student student) {
+    ResponseEntity<String> addStudent(@RequestBody Student student) {
+        if (student.getId() < 0 || student.getAge() < 6) {
+            return ResponseEntity.badRequest().header("Successful:", "false").body("Invalid id or age");
+        }
+
         studentService.addStudent(student);
+        return ResponseEntity.ok().header("Successful:", "true").body("Student added");
     }
 
     @PutMapping(path = "{id}")
-    public void editStudent(@PathVariable("id") Long id, @RequestBody Student student) {
+    ResponseEntity<String> editStudent(@PathVariable("id") Long id, @RequestBody Student student) {
+        if (student.getId() < 0 || student.getAge() < 6 || id < 0) {
+            return ResponseEntity.badRequest().header("Successful:", "false").body("Invalid id or age");
+        }
+
         studentService.editStudent(id, student);
+        return ResponseEntity.ok().header("Successful:", "true").body("Student updated");
     }
 
     @DeleteMapping(path = "{id}")
-    public void deleteStudent(@PathVariable("id") Long id) {
+    ResponseEntity<String> deleteStudent(@PathVariable("id") Long id) {
+        if (id < 0) {
+            return ResponseEntity.badRequest().header("Successful:", "false").body("Invalid id or age");
+        }
+
         studentService.deleteStudent(id);
+        return ResponseEntity.ok().header("Successful:", "true").body("Student deleted");
     }
 
     @GetMapping(path = "{id}")
-    public Student getStudent(@PathVariable("id") Long id) {
-        return studentService.getStudent(id);
+    ResponseEntity<String> getStudent(@PathVariable("id") Long id) {
+        if (id < 0) {
+            return ResponseEntity.badRequest().header("Successful:", "false").body("Invalid id or age");
+        }
+
+        return ResponseEntity.ok().header("Successful:", "true")
+                        .body(SpringRestApiLbdApplication.asJsonString(studentService.getStudent(id)));
     }
 }
